moves PersonalData from old folder structure to new
